# services/aiconfigurator/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AIConfigurator API
  description: |
    Minimal REST API for LLM inference GPU sizing and memory estimation.
    Wraps the aiconfigurator SDK (`cli_recommend` and `estimate_kv_cache`).
  version: 1.0.0
  contact:
    name: ConfigIQ team
servers:
  - url: http://localhost:7860
    description: Local development
  - url: https://aiconfigurator.dev
    description: Production
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [catalog]
      summary: List supported models
      description: |
        Returns supported model identifiers. Without include, returns a sorted list of strings.
        Use include=specs to get per-model metadata (MoE fields, context length, attention heads, architecture).
      parameters:
        - name: include
          in: query
          required: false
          description: Comma-separated extras to include (specs).
          schema:
            type: string
            example: "specs"
      responses:
        '200':
          description: Model list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
  /systems:
    get:
      operationId: listSystems
      tags: [catalog]
      summary: List supported GPU systems
      description: |
        Returns supported GPU systems. Each entry always includes id and name.
        Use include=specs to add hardware details (memory, TDP, architecture).
      parameters:
        - name: include
          in: query
          required: false
          description: Comma-separated extras to include (specs).
          schema:
            type: string
            example: "specs"
      responses:
        '200':
          description: System list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemListResponse'
...
components:
  schemas:
...
    ModelListResponse:
      type: object
      properties:
        models:
          type: array
          description: Sorted list of models. Items are strings without include=specs, objects with include=specs.
          items:
            oneOf:
              - type: string
              - $ref: '#/components/schemas/ModelEntry'
    ModelEntry:
      type: object
      required: [id]
      properties:
        id:
          type: string
          description: HuggingFace model identifier.
          example: Qwen/Qwen3-235B-A22B
        num_experts:
          type: integer
          nullable: true
          description: Total number of experts (null for dense models).
        num_experts_per_tok:
          type: integer
          nullable: true
          description: Experts activated per token — top-k (null for dense models).
        context_length:
          type: integer
          nullable: true
          description: Maximum context window in tokens.
        num_attn_heads:
          type: integer
          nullable: true
          description: Number of query attention heads.
        num_kv_heads:
          type: integer
          nullable: true
          description: Number of KV heads (less than num_attn_heads when GQA is used).
        architecture:
          type: string
          nullable: true
          description: HuggingFace model architecture class name.
          example: Qwen3MoeForCausalLM
    SystemListResponse:
      type: object
      properties:
        systems:
          type: array
          items:
            $ref: '#/components/schemas/SystemEntry'
          description: List of supported GPU systems.
    SystemEntry:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          description: System identifier.
          example: h200_sxm
        name:
          type: string
          description: Human-readable GPU name.
          example: NVIDIA H200 SXM
        # Fields below are present when include=specs
        vendor:
          type: string
          nullable: true
          example: nvidia
        architecture:
          type: string
          nullable: true
          example: hopper
        memory_bytes:
          type: integer
          description: GPU memory capacity in bytes.
        memory_bandwidth_bytes:
          type: integer
          description: Peak memory bandwidth in bytes per second.
        bf16_tflops:
          type: number
          description: BF16 tensor core throughput in TFLOPS.
        tdp_watts:
          type: number
          description: Thermal design power (watts).
        gpus_per_node:
          type: integer
          description: Number of GPUs per node.

    # ── Error ──────────────────────────────────────────────────────────────
...